pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "staticsite/__init__.py"

[project]
name = "staticsite"
dynamic = ["version"]
authors = [
  {name = "Enrico Zini", email = "enrico@enricozini.org"}
]
description = "Static site generator"
readme = "README.md"
requires-python = ">= 3.11"
license = {file = "LICENSE"}
classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet :: WWW/HTTP :: Site Management",
]
dependencies = [
  'markdown',
  'docutils',
  'toml',
  'pyyaml',
  'ruamel.yaml',
  'jinja2>3',
  'Pillow',
  'python_dateutil',
  'python_slugify',
  'pytz'
]

[project.optional-dependencies]
serve = ["tornado", "pyinotify"]
fast_caching = ["lmdb"]

[project.scripts]
ssite = "staticsite.__main__:run_main"

[project.urls]
homepage = "https://github.com/spanezz/staticsite"

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html
mypy_path = "stubs"

# Warn about casting an expression to its inferred type (inverse: --no-warn-redundant-casts)
warn_redundant_casts = true

# Warn about unneeded '# type: ignore' comments (inverse: --no-warn-unused-ignores)
warn_unused_ignores = true

# Warn about returning values of type Any from non-Any typed functions (inverse: --no-warn-return-any)
warn_return_any = true

#  --no-warn-no-return       Do not warn about functions that end without returning (inverse: --warn-no-return)

# Warn about statements or expressions inferred to be unreachable (inverse: --no-warn-unreachable)
warn_unreachable = true

# Disallow defining functions without type annotations or with incomplete type annotations (inverse: --allow-untyped-defs)
disallow_untyped_defs = true

# Disallow defining functions with incomplete type annotations (inverse: --allow-incomplete-defs)
disallow_incomplete_defs = true

# Type check the interior of functions without type annotations (inverse: --no-check-untyped-defs)
check_untyped_defs = true

# Treat imports as private unless aliased (inverse: --implicit-reexport)
no_implicit_reexport = true

# Disallow decorating typed functions with untyped decorators (inverse: --allow-untyped-decorators)
disallow_untyped_decorators = true

# Strict mode; enables the following flags: --warn-unused-configs, --disallow-any-generics, --disallow-subclassing-any, --disallow-untyped-calls, --disallow-untyped-defs,
# --disallow-incomplete-defs, --check-untyped-defs, --disallow-untyped-decorators, --warn-redundant-casts, --warn-unused-ignores, --warn-return-any, --no-implicit-
# reexport, --strict-equality, --strict-concatenate
strict = true

# [[tool.mypy.overrides]]
# module = ["staticsite.site", "staticsite.page", "staticsite.theme"]
# 
# #  --disallow-any-unimported
# #                            Disallow Any types resulting from unfollowed imports
# #  --disallow-any-expr       Disallow all expressions that have type Any
# #  --disallow-any-decorated  Disallow functions that have Any in their signature after decorator transformation
# #  --disallow-any-explicit   Disallow explicit Any in type positions
# #  --disallow-any-generics   Disallow usage of generic types that do not specify explicit type parameters (inverse: --allow-any-generics)
# #  --disallow-subclassing-any
# #                            Disallow subclassing values of type 'Any' when defining classes (inverse: --allow-subclassing-any)
# #
# #  --disallow-untyped-calls  Disallow calling functions without type annotations from functions with type annotations (inverse: --allow-untyped-calls)
# #  --allow-untyped-globals   Suppress toplevel errors caused by missing annotations (inverse: --disallow-untyped-globals)
# #  --allow-redefinition      Allow unconditional variable redefinition with a new type (inverse: --disallow-redefinition)
# #  --strict-equality         Prohibit equality, identity, and container checks for non-overlapping types (inverse: --no-strict-equality)
# #  --strict-concatenate      Make arguments prepended via Concatenate be truly positional-only (inverse: --no-strict-concatenate)
